requirements.txt
httpx==0.26.0
fastapi==0.108.0
uvicorn==0.25.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0